Fix stuff

+++ server/inputpin.h
@@ -8,12 +8,12 @@
 class InputPin : public Pin
 {
     public:
-        InputPin(int pinNumber, int pullUp);
-        void set(int data);
-        int get();
+        InputPin(uint8_t pinNumber, uint8_t pullUp);
+        void set(uint8_t data);
+        uint8_t get();
 
     private:
-        int pullUp;
+        uint8_t pullUp;
 };
 
 #endif // INPUTPIN_H
